+++ .gitignore
@@ -22,3 +22,5 @@ yarn-error.log*
 
 # backups
 *~
+
+/.vscode
