+++ src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './styles/App.css';
+import styles from './styles/App.css';  // eslint-disable-line
 import SimulationMenu from './components/SimulationMenu.js';
 
 export default function App() {
