+++ src/lib/data-generation.js
@@ -1,0 +1,144 @@
+import chi from 'chi-squared';
+
+export const generateNormal = (sampleSize) => {
+  const MEAN = 64;
+  const STANDARD_DEV = 3;
+  const ITERATES = 9;
+  const range = Math.sqrt(12) * STANDARD_DEV * STANDARD_DEV;
+  const popMin = MEAN - (range / 2);
+
+  const popArray = [];
+
+  let dict = Array(sampleSize).fill(-1);
+
+  // creates data points for population and stores it in popArray
+  for (let i = 0; i < sampleSize; i++){
+      let sum = 0;
+      for (let j = 0; j < ITERATES; j++){
+          sum += Math.random() * range + popMin;
+      }
+
+      if (dict[Math.round(sum / ITERATES * 10)] !== -1){
+          dict[Math.round(sum / ITERATES * 10)] += 1;
+      }
+      // Adds first instance of a point
+      else {
+          dict[Math.round(sum / ITERATES * 10)] = 1;
+      }
+  }
+
+  for (const point in dict) {
+      if (point !== -1) {
+          for (let count = 1; count < dict[point] + 1; count++) {
+              popArray.push([point/10, count]);
+          }
+      }
+  }
+  popArray.sort(() => Math.random() - 0.5);
+  popArray.sort((a,b) => b[1] - a[1]);
+
+  return popArray;
+}
+
+export const generateUniform = (mainSampleSize) => {
+  const HI = 10;
+  const LOW = -10;
+  const range = HI - LOW;
+
+  const popArr = []
+
+  const sampleSize = mainSampleSize - 100;
+  let dict = Array(sampleSize).fill(-1);
+
+  for (let i = 0; i < sampleSize; i++){
+      const val = Math.random() * range + LOW;
+
+      if (dict[Math.round(val * 10)]){
+          dict[Math.round(val * 10)] += 1;
+      } else {
+          dict[Math.round(val * 10)] = 1;
+      }
+  }
+
+  for (const point in dict) {
+      if (point !== -1) {
+          for (let count = 1; count < dict[point] + 2; count++) {
+              popArr.push([point/10, count]);
+          }
+      }
+  }
+
+  popArr.sort(() => Math.random() - 0.3);
+  popArr.sort((a,b) => b[1] - a[1]);
+
+  return popArr;
+}
+
+export const generateExponential = (sampleSize) => {
+  const LAMBDA = 1/64;
+
+  const popArray =  [];
+
+  let dict = Array(sampleSize).fill(-1);
+
+  for (let i = 0; i < sampleSize; i++){
+      const val = -Math.log(Math.random()) / LAMBDA
+
+      if (dict[Math.round(val * 10)]){
+          dict[Math.round(val * 10)] += 1;
+      } else {
+          dict[Math.round(val * 10)] = 1;
+      }
+  }
+
+  for (const point in dict) {
+      if (point !== -1) {
+          for (let count = 1; count < dict[point] + 2; count++) {
+              popArray.push([point/10, count]);
+          }
+      }
+  }
+  popArray.sort(() => Math.random() - 0.5);
+  popArray.sort((a,b) => b[1] - a[1]);
+  return popArray;
+}
+
+
+export const generateChiSquared = (sampleSize) => {
+  const DEGREES_OF_FREEDOM = 8;
+  const chiArray = [];
+  const chiMin = chi.pdf(20, DEGREES_OF_FREEDOM);
+  for (let i = 0; i < 20; i+=.1){
+      const tmp = chi.pdf(i, DEGREES_OF_FREEDOM)
+      for (let j = 0; j < tmp / chiMin; j++){
+          chiArray.push(i)
+      }
+  }
+
+  const popArray = [];
+
+  let dict = Array(sampleSize).fill(-1);
+
+  for (let i = 0; i < sampleSize; i++){
+      const val = chiArray[Math.round(Math.random() * chiArray.length)];
+
+      if (dict[Math.round(val * 10)]){
+          dict[Math.round(val * 10)] += 1;
+      } else {
+          dict[Math.round(val * 10)] = 1;
+      }
+  }
+
+  for (const point in dict) {
+      if (point !== -1) {
+          for (let count = 1; count < dict[point] + 2; count++) {
+              popArray.push([point/10, count]);
+          }
+      }
+  }
+
+  popArray.sort(() => Math.random() - 0.5);
+  popArray.sort((a,b) => b[1] - a[1]);
+
+  return popArray;
+}
